Remove TTS function definitions and related tests
